+++ core/server/api_container/server/api_container_service.go
@@ -161,19 +161,6 @@ func (apicService ApiContainerService) ExecuteModule(ctx context.Context, args *
 	return resp, nil
 }
 
-func (apicService ApiContainerService) RegisterService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.RegisterServiceArgs) (*kurtosis_core_rpc_api_bindings.RegisterServiceResponse, error) {
-	serviceId := kurtosis_backend_service.ServiceID(args.ServiceId)
-	partitionId := service_network_types.PartitionID(args.PartitionId)
-
-	privateIpAddr, err := apicService.serviceNetwork.RegisterService(ctx, serviceId, partitionId)
-	if err != nil {
-		// TODO IP: Leaks internal information about API container
-		return nil, stacktrace.Propagate(err, "An error occurred registering service '%v' in the service network", serviceId)
-	}
-
-	return binding_constructors.NewRegisterServiceResponse(privateIpAddr.String()), nil
-}
-
 func (apicService ApiContainerService) RegisterServices(ctx context.Context, args *kurtosis_core_rpc_api_bindings.RegisterServicesArgs) (*kurtosis_core_rpc_api_bindings.RegisterServicesResponse, error) {
 	serviceIDs := map[kurtosis_backend_service.ServiceID]bool{}
 	for id := range args.ServiceIdSet {
@@ -192,91 +179,6 @@ func (apicService ApiContainerService) RegisterServices(ctx context.Context, arg
 	}
 
 	return binding_constructors.NewRegisterServicesResponse(serviceIDsToIPsStringMap), nil
-}
-
-func (apicService ApiContainerService) StartService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.StartServiceArgs) (*kurtosis_core_rpc_api_bindings.StartServiceResponse, error) {
-	logrus.Debugf("Received request to start service with the following args: %+v", args)
-	serviceId := kurtosis_backend_service.ServiceID(args.ServiceId)
-	privateApiPorts := args.PrivatePorts
-
-	//TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
-	requestedPublicApiPorts := args.PublicPorts
-	if len(requestedPublicApiPorts) > 0 {
-
-		if len(privateApiPorts) != len(requestedPublicApiPorts) {
-			return nil, stacktrace.NewError("The received private ports length and the public ports length are not equal, received '%v' private ports and '%v' public ports", len(privateApiPorts), len(requestedPublicApiPorts))
-		}
-
-		for portId, privatePort := range privateApiPorts {
-			if _, found := requestedPublicApiPorts[portId]; !found {
-				return nil, stacktrace.NewError("Expected to receive public port with ID '%v' bound to private port number '%v', but it was not found", portId, privatePort.GetNumber())
-			}
-		}
-	}
-
-	requestedPublicServicePortSpecs := map[string]*port_spec.PortSpec{}
-	for portId, publicApiPort := range requestedPublicApiPorts {
-		publicServicePortSpec, err := transformApiPortToPortSpec(publicApiPort)
-		if err != nil {
-			return nil, stacktrace.NewError("An error occurred transforming the API port for public port '%v' into a port spec port", portId)
-		}
-		requestedPublicServicePortSpecs[portId] = publicServicePortSpec
-	}
-	//TODO Finished the huge hack to temporarily enable static ports for NEAR
-
-	privateServicePortSpecs := map[string]*port_spec.PortSpec{}
-	for portId, privateApiPort := range privateApiPorts {
-		privateServicePortSpec, err := transformApiPortToPortSpec(privateApiPort)
-		if err != nil {
-			return nil, stacktrace.NewError("An error occurred transforming the API port for private port '%v' into a port spec port", portId)
-		}
-		privateServicePortSpecs[portId] = privateServicePortSpec
-	}
-	filesArtifactMountpointsByArtifactUuid := map[enclave_data_directory.FilesArtifactUUID]string{}
-	for filesArtifactUuidStr, mountDirPath := range args.FilesArtifactMountpoints {
-		filesArtifactMountpointsByArtifactUuid[enclave_data_directory.FilesArtifactUUID(filesArtifactUuidStr)] = mountDirPath
-	}
-	startedService, err := apicService.serviceNetwork.StartService(
-		ctx,
-		serviceId,
-		args.DockerImage,
-		privateServicePortSpecs,
-		requestedPublicServicePortSpecs,
-		args.EntrypointArgs,
-		args.CmdArgs,
-		args.DockerEnvVars,
-		filesArtifactMountpointsByArtifactUuid,
-		args.CpuAllocationMillicpus,
-		args.MemoryAllocationMegabytes,
-	)
-	if err != nil {
-		// TODO IP: Leaks internal information about the API container
-		return nil, stacktrace.Propagate(err, "An error occurred starting the service in the service network")
-	}
-	privateServiceIpStr := startedService.GetRegistration().GetPrivateIP().String()
-	serviceGuidStr := string(startedService.GetRegistration().GetGUID())
-	publicServicePortSpecs := startedService.GetMaybePublicPorts()
-	publicApiPorts, err := transformPortSpecMapToApiPortsMap(publicServicePortSpecs)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred transforming the service's public port specs to API ports")
-	}
-	maybePublicIpAddr := startedService.GetMaybePublicIP()
-	publicIpAddrStr := missingPublicIpAddrStr
-	if maybePublicIpAddr != nil {
-		publicIpAddrStr = maybePublicIpAddr.String()
-	}
-	response := binding_constructors.NewStartServiceResponse(privateServiceIpStr, privateApiPorts, publicIpAddrStr, publicApiPorts, serviceGuidStr)
-
-	serviceStartLoglineSuffix := ""
-	if len(publicServicePortSpecs) > 0 {
-		serviceStartLoglineSuffix = fmt.Sprintf(
-			" with the following public ports: %+v",
-			publicServicePortSpecs,
-		)
-	}
-	logrus.Infof("Started service '%v'%v", serviceId, serviceStartLoglineSuffix)
-
-	return response, nil
 }
 
 func (apicService ApiContainerService) StartServices(ctx context.Context, args *kurtosis_core_rpc_api_bindings.StartServicesArgs) (*kurtosis_core_rpc_api_bindings.StartServicesResponse, error){